feat(dep): use single-digit dependency notation
Use `^3` instead of `^3.1.2` cause it’s more simple and requires fewer changes. Yarn doesn’t update the package.json unless you use specific version numbers or `--latest` on upgrade command. So you can switch to single-digit notation in your package.json and yarn won’t touch it.

Using single digits, you don’t have to worry if your lockfile is out of sync with the package.json. Of course it is, and now it’s more obvious! Accept yarn.lock as single point of truth.

Note that this workflow is not common practise but sort of a field test! :)